src/urscript/auto-backup-script/requirements.txt
# Python dependencies for UR3 Auto-Backup Script
paho-mqtt>=1.6.0
schedule>=1.2.0
